feat: move schemas and actions into lib folder
