+++ lib/schemas/index.ts
@@ -70,3 +70,30 @@ export const RegisterSchema = z
     message: 'Passwords must match.',
     path: ['confirmPassword'],
   });
+
+export const CourseCreationSchema = z.object({
+  term: z.string({
+    required_error: 'Please select a term offering.',
+  }),
+  code: z
+    .string()
+    .length(8, { message: 'Course code must be exactly 8 characters.' })
+    .refine((code) => /^[A-Za-z]{4}\d{4}$/.test(code), {
+      message: 'Course code must begin with 4 letters and end with 4 numbers.',
+    }),
+  name: z
+    .string()
+    .max(120, {
+      message: 'Course name must not exceed 120 characters.',
+    })
+    .nonempty('Course name is required.'),
+  // faculty: z.string({
+  //   required_error: 'Please select a faculty.',
+  // }),
+  description: z
+    .string()
+    .max(200, {
+      message: 'Course description must not exceed 200 characters.',
+    })
+    .nonempty('Course description is required.'),
+});
